fixed autocompletion in free-coding page
