+++ PlaygroundBook/Chapters/Chapter5.playgroundchapter/Pages/Page1.playgroundpage/Contents.swift
@@ -5,18 +5,23 @@ import Foundation
 import PlaygroundSupport
 //#-end-hidden-code
 
+//#-code-completion(bookauxiliarymodule, hide)
+//#-code-completion(chapterauxiliarymodule, show)
+//#-code-completion(currentmodule, show)
+//#-code-completion(identifier, show, A, B, AB)
+//#-code-completion(identifier, show, red, green, blue, yellow, black, darkGray, lightGray, white, cyan, magenta, orange, purple)
+//#-code-completion(identifier, show, miniImage, smiley, sad, heart, arrow_left, arrow_right, arrow_left_right, full, dot, small_rect, large_rect, double_row, tick, rock, scissors, well, flash, wave)
+//#-code-completion(identifier, show, C, D, E, F, G, A, H, C5)
+
 //#-hidden-code
 playgroundPrologue()
 //#-end-hidden-code
 
 class CalliopeImplementation: Calliope {
-	//#-editable-code #localized(key: "cc.calliope.miniplaygroundbook.Advanced.EditableCalliope")
+	//#-editable-code cc.calliope.miniplaygroundbook.Advanced.EditableCalliope
 	//#-end-editable-code
 }
 
 //#-hidden-code
 myCalliope = CalliopeImplementation()
 //#-end-hidden-code
-
-//#-editable-code #localized(key: "cc.calliope.miniplaygroundbook.Advanced.MainEditableArea")
-//#-end-editable-code
