+++ PlaygroundBook/PrivateResources/de.lproj/EditableFields.strings
@@ -3,3 +3,5 @@
 "Tap to write your code" = "Tippe um deinen Code zu schreiben";
 
 "Choose a number" = "Wähle eine Zahl";
+
+"cc.calliope.miniplaygroundbook.Advanced.EditableCalliope" = "Implementiere die Methoden deines Calliope";
