+++ PlaygroundBook/PrivateResources/en.lproj/EditableFields.strings
@@ -2,3 +2,5 @@
 "Tap to write your code" = "Tap to write your code";
 
 "Choose a number" = "Choose a number";
+
+"cc.calliope.miniplaygroundbook.Advanced.EditableCalliope" = "Implement the methods of your callipe";
