+++ PlaygroundBook/Sources/LiveView/Auxiliary/Extension_CGRect.swift
@@ -2,7 +2,7 @@ import Foundation
 import UIKit
 
 extension CGRect {
-    public func insetBy(percentage: CGFloat) -> CGRect {
+    func insetBy(percentage: CGFloat) -> CGRect {
         let startWidth = self.width
         let startHeight = self.height
         let adjustmentWidth = (startWidth-(startWidth * percentage)) * 0.5
