--- Playground_Code/Playground_Code.swift
@@ -1,10 +1,0 @@
-//
-//  Playground_Code.swift
-//  Playground_Code
-//
-//  Created by Tassilo Karge on 27.01.19.
-//
-
-class Playground_Code {
-
-}
