--- libPlayground_Code/libPlayground_Code.swift
@@ -1,10 +1,0 @@
-//
-//  libPlayground_Code.swift
-//  libPlayground_Code
-//
-//  Created by Tassilo Karge on 27.01.19.
-//
-
-class libPlayground_Code {
-
-}
